#01 setting: Application/MainActivity 설정
