+++ app/src/main/AndroidManifest.xml
@@ -11,11 +11,11 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.IEUM"
+        android:name=".IEUMApplication"
         tools:targetApi="31" >
         <activity
-            android:name=".MainActivity"
+            android:name="com.ieum.presentation.MainActivity"
             android:exported="true"
-            android:label="@string/app_name"
             android:theme="@style/Theme.IEUM" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
