+++ app/src/main/java/com/ieum/android/IEUMApplication.kt
@@ -1,0 +1,5 @@
+package com.ieum.android
+
+import android.app.Application
+
+class IEUMApplication : Application()
